html/parsing/html5lib/html-tree/isindex.dat

#data
<isindex>
#errors
no DOCTYPE
isindex
#document
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <label>
|         "This is a searchable index. Enter search keywords: "
|         <input>
|           name="isindex"
|       <hr>

#data
<isindex name="A" action="B" prompt="C" foo="D">
#errors
no DOCTYPE
isindex
#document
| <html>
|   <head>
|   <body>
|     <form>
|       action="B"
|       <hr>
|       <label>
|         "C"
|         <input>
|           foo="D"
|           name="isindex"
|       <hr>

#data
<form><isindex>
#errors
no DOCTYPE
isindex
EOF
#document
| <html>
|   <head>
|   <body>
|     <form>
